PythonProject/requirements.txt
numpy==1.13.3
pip==9.0.1
PySDL2==0.9.6
setuptools==28.8.0
